Test1-Django/requirements.txt
asgiref==3.7.2
basicauth==0.4.1
certifi==2023.7.22
charset-normalizer==3.3.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.14.0
gunicorn==21.2.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
Pillow==10.0.1
pre-commit
psycopg2-binary==2.9.8
pycryptodome==3.15.0
PyJWT==2.5.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.6
pre-commit
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0